Move packageInfo to utils
